Split QuestHandler into PlayerQuestHandler and GlobalQuestHandler. PlayerQuestHandler handles data specific to each player, while GlobalQuestHandler handles data specific to quests themselves.
VERSION fields must now start at 1 because NBT.getInt() defaults to 0 if
it cannot find the value, and we must determine if the value exists.

Quests, rewards, and tasks will now require a MinecraftServer instance to
be instantiated, so we have access to the world at all times.

Added an NBT debugger to help with debugging the state of NBT.